Move snippet viewsets registration to register_snippet()
