+++ wagtail/snippets/wagtail_hooks.py
@@ -2,7 +2,6 @@ from django.contrib.admin.utils import quote
 from django.contrib.auth.models import Permission
 from django.contrib.contenttypes.models import ContentType
 from django.urls import include, path, reverse
-from django.utils.module_loading import import_string
 from django.utils.translation import gettext_lazy as _
 
 from wagtail import hooks
@@ -14,7 +13,6 @@ from wagtail.snippets.permissions import (
     user_can_edit_snippet_type,
     user_can_edit_snippets,
 )
-from wagtail.snippets.views import chooser as chooser_views
 from wagtail.snippets.views import snippets as snippet_views
 from wagtail.snippets.widgets import SnippetListingButton
 
@@ -31,31 +29,6 @@ def register_admin_urls():
     return [
         path("snippets/", include(snippet_index_patterns)),
     ]
-
-
-@hooks.register("register_admin_viewset")
-def register_viewsets():
-    viewsets = []
-    for model in get_snippet_models():
-        admin_viewset = (
-            getattr(model, "admin_viewset", None) or snippet_views.SnippetViewSet
-        )
-        if isinstance(admin_viewset, str):
-            admin_viewset = import_string(admin_viewset)
-
-        viewsets += [
-            admin_viewset(
-                model.get_admin_url_namespace(),
-                model=model,
-                url_prefix=model.get_admin_base_path(),
-            ),
-            chooser_views.SnippetChooserViewSet(
-                f"wagtailsnippetchoosers_{model._meta.app_label}_{model._meta.model_name}",
-                model=model,
-                url_prefix=f"snippets/choose/{model._meta.app_label}/{model._meta.model_name}",
-            ),
-        ]
-    return viewsets
 
 
 class SnippetsMenuItem(MenuItem):
